This gets working with ssh and two snapshots.  The primary change was to have backup_handler() be synchronous, since nothing was waiting for it and cleaning up. (And this resulted in the dataset being busy on the remote end.)
Also turn all my debugging prints back to if debug.

Not finished with this branch yet:  I need to make error checking
actually work, and backup_handler() should probably still be
asynchronous, but I need to add some code for backup() to wait for
it, along with all the filters.